Ruby 3.2 port.  READ AI CONTENT CAUTION BELOW:
The differences between Ruby 2 and Ruby 3 are significant
in terms of thread safety and memory safety.  I (Cal) have not
worked with the Ruby SDK very much in the past.  There
is not much documentation for C++ embedders to use, especially
concerning the newer Ruby versions.

So I decided to try out OpenAI's Codex CLI.  I gave it access
to all of Ruby 3.2's source and headers, along with a mechanism
to compile and test ModRuby on a live Apache server.

I showed it the segfault back trace in rb_protect() and told it
to resolve the issue.  It spent a total of around an hour
doing web searches, reading the source and trying things out.
It made mistakes, spotted its own errors, then recompiled and retested
until it had success.

I've scanned over this commit and I don't see any obvious problems
with it.

There were two main issues here for the Ruby 3 port:

1) You can't call rb_protect() without a global VM lock
2) You can't get the global VM lock from a thread outside of Ruby
   (IE:  In an Apache MPM thread)

Fixing #1 caused #2.  The RubyDispatcher class shifts interaction
with Ruby to an interface that manages the global VM.  rb_protect()
is wrapped in a GVL so that native C++ functions can interact
with Ruby objects and methods.